refactor: Extract connect
